+++ src/wildboar/metrics/__init__.py
@@ -1,5 +1,6 @@
 """Evaluation metrics."""
 
+from ._cluster import silhouette_samples, silhouette_score
 from ._counterfactual import (
     compactness_score,
     plausability_score,
@@ -10,10 +11,14 @@ from ._counterfactual import (
 )
 
 __all__ = [
+    # Counterfactuals
     "proximity_score",
     "relative_proximity_score",
     "compactness_score",
     "validity_score",
     "plausability_score",
     "redudancy_score",
+    # Cluster
+    "silhouette_samples",
+    "silhouette_score",
 ]